+++ web/src/admin/users/UserApplicationTable.ts
@@ -1,85 +1,21 @@
+import "#admin/users/AgentAddApplicationForm";
 import "#elements/AppIcon";
-import "#elements/forms/HorizontalFormElement";
-import "#elements/forms/SearchSelect/index";
 import "@patternfly/elements/pf-tooltip/pf-tooltip.js";
 
 import { DEFAULT_CONFIG } from "#common/api/config";
 
 import { renderModal } from "#elements/dialogs";
-import { Form } from "#elements/forms/Form";
 import { PaginatedResponse, Table, TableColumn } from "#elements/table/Table";
 import { SlottedTemplateResult } from "#elements/types";
 import { ifPresent } from "#elements/utils/attributes";
 
 import { applicationListStyle } from "#admin/applications/ApplicationListPage";
 
-import {
-    Application,
-    CoreApi,
-    CoreApplicationsListRequest,
-    User,
-    UserTypeEnum,
-} from "@goauthentik/api";
+import { Application, CoreApi, User, UserTypeEnum } from "@goauthentik/api";
 
 import { msg } from "@lit/localize";
-import { CSSResult, TemplateResult, html, nothing } from "lit";
+import { CSSResult, html, nothing } from "lit";
 import { customElement, property } from "lit/decorators.js";
-
-const USER_ATTRIBUTE_AGENT_OWNER_PK = "goauthentik.io/agent/owner-pk";
-
-@customElement("ak-agent-add-application-form")
-export class AgentAddApplicationForm extends Form<{ app: string }> {
-    public override headline = msg("Add Application");
-    public override submitLabel = msg("Add");
-
-    @property({ attribute: false })
-    public agent: User | null = null;
-
-    public override getSuccessMessage(): string {
-        return msg("Successfully added application.");
-    }
-
-    async send(data: { app: string }): Promise<{ app: string }> {
-        if (!this.agent) throw new Error("Agent not set");
-        await new CoreApi(DEFAULT_CONFIG).coreUsersAgentAllowedAppPartialUpdate({
-            id: this.agent.pk,
-            patchedUserAgentAllowedAppRequest: { app: data.app, action: "add" },
-        });
-        return data;
-    }
-
-    protected override renderForm(): TemplateResult {
-        const ownerPk = this.agent?.attributes?.[USER_ATTRIBUTE_AGENT_OWNER_PK];
-
-        return html`<ak-form-element-horizontal
-                label=${msg("Application")}
-                required
-                name="app"
-            >
-                <ak-search-select
-                    placeholder=${msg("Select an application...")}
-                    .fetchObjects=${async (query?: string): Promise<Application[]> => {
-                        const args: CoreApplicationsListRequest = {
-                            ordering: "name",
-                            pageSize: 20,
-                            forUser: ownerPk ? Number(ownerPk) : undefined,
-                        };
-                        if (query) {
-                            args.search = query;
-                        }
-                        const result =
-                            await new CoreApi(DEFAULT_CONFIG).coreApplicationsList(args);
-                        return result.results;
-                    }}
-                    .renderElement=${(app: Application): string => app.name}
-                    .value=${(app: Application | undefined): string | undefined => app?.pk}
-                    .renderDescription=${(app: Application): TemplateResult =>
-                        html`${app.group || msg("No group")}`}
-                >
-                </ak-search-select>
-            </ak-form-element-horizontal>`;
-    }
-}
 
 @customElement("ak-user-application-table")
 export class UserApplicationTable extends Table<Application> {
@@ -118,11 +54,13 @@ export class UserApplicationTable extends Table<Application> {
     }
 
     protected openAddApplicationModal = () => {
-        return renderModal(
+        renderModal(
             html`<ak-agent-add-application-form
                 .agent=${this.user}
             ></ak-agent-add-application-form>`,
-        );
+        ).then(() => {
+            this.fetch();
+        });
     };
 
     protected override renderToolbar(): SlottedTemplateResult {
@@ -189,6 +127,5 @@ export class UserApplicationTable extends Table<Application> {
 declare global {
     interface HTMLElementTagNameMap {
         "ak-user-application-table": UserApplicationTable;
-        "ak-agent-add-application-form": AgentAddApplicationForm;
     }
 }
